Add getAll() and getByName() class methods to P100
These methods log in to the tplink cloud to get a list of MAC addresses for relevant devices, then send a broadcast packet to see if those devices respond, and then either instantiate all the objects or just the one with the correct alias and return it.
The P100 class now also has a shared UUID for api calls that stays the same from when the class is imported
Example uses of getAll() and getByName() have been added to README.md
Additionally I changed _generateException to be a static method so it can be used by the factory class methods added in this commit.